chore: moved all helpers and types to diff files
