+++ app/client/src/workers/DataTreeEvaluator/index.ts
@@ -63,6 +63,7 @@ import {
   Severity,
   SourceEntity,
   ENTITY_TYPE as CONSOLE_ENTITY_TYPE,
+  UserLogObject,
 } from "entities/AppsmithConsole";
 import { error as logError } from "loglevel";
 import { JSUpdate } from "utils/JSPaneUtils";
@@ -83,7 +84,6 @@ import {
   parseJSActions,
 } from "workers/JSObject";
 import { lintTree } from "workers/Lint";
-import { UserLogObject } from "workers/UserLog";
 
 export default class DataTreeEvaluator {
   dependencyMap: DependencyMap = {};
