+++ app/client/src/workers/evaluation.worker.ts
@@ -26,7 +26,7 @@ import { setFormEvaluationSaga } from "./formEval";
 import { isEmpty } from "lodash";
 import { EvalMetaUpdates } from "./DataTreeEvaluator/types";
 import { EvalTreePayload } from "../sagas/EvaluationsSaga";
-import { UserLogObject } from "./UserLog";
+import { UserLogObject } from "entities/AppsmithConsole";
 
 const CANVAS = "canvas";
